Refactored controller to VideogameController
