+++ hello-grandma/src/main/java/com/mindera/controllers/VideogameController.java
@@ -1,32 +1,58 @@
 package com.mindera.controllers;
 
-import com.mindera.services.implementations.VideogameServiceImpl;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.mindera.entities.Videogame;
+import com.mindera.exceptions.videogame.VideogameNotFoundException;
+import com.mindera.repositories.VideogameExtensionRepository;
+import com.mindera.services.implementations.VideogameExtensionServiceImpl;
 import jakarta.inject.Inject;
 import jakarta.ws.rs.*;
 import jakarta.ws.rs.core.MediaType;
-/*import org.jboss.resteasy.reactive.RestResponse;*/
+import org.eclipse.microprofile.rest.client.inject.RegisterRestClient;
+
+import java.util.List;
 
 @Path("/api/v1/videogames")
-@Produces(MediaType.APPLICATION_JSON)
-@Consumes(MediaType.APPLICATION_JSON)
-public class VideogameController {
+@RegisterRestClient(configKey="igdb-api")
+public class VideogameController implements VideogameExtensionRepository {
 
     @Inject
-    VideogameServiceImpl videogameServiceImpl;
+    VideogameExtensionServiceImpl igdbService;
 
-    /*@GET
-    public RestResponse<List<VideogameGetDto>> getAllVideogames() {
-        return RestResponse.ok(videogameService.getAllVideogames());
-    }
 
     @GET
-    @Path("/{id}")
-    public RestResponse<VideogameGetDto> getVideogameById(@PathParam("id") String id) throws VideogameNotFoundException, VideogameNotFoundException {
-        return RestResponse.ok(videogameService.findById(id));
+    @Path("/videogames")
+    @Produces(MediaType.APPLICATION_JSON)
+    public List<Videogame> getAll() throws VideogameNotFoundException, JsonProcessingException {
+        return igdbService.getAll();
+    }
+
+
+    @POST
+    @Path("/videogames")
+    @Produces(MediaType.APPLICATION_JSON)
+    public List<Videogame> getAllVideogames() throws VideogameNotFoundException, JsonProcessingException {
+        return igdbService.getAllVideogames();
     }
 
     @POST
-    public RestResponse<VideogameGetDto> create(VideogameCreateDto videogame) {
-        return RestResponse.accepted(videogameService.create(videogame));
-    }*/
+    @Path("/videogames/{id}")
+    @Produces(MediaType.APPLICATION_JSON)
+    public Videogame findById(@PathParam("id") String id) throws VideogameNotFoundException, JsonProcessingException {
+        return igdbService.findById(Integer.parseInt(id));
+    }
+
+    @POST
+    @Path("/search/{search}")
+    @Produces(MediaType.APPLICATION_JSON)
+    public List<Videogame> findBySearch(@PathParam("search") String search) throws VideogameNotFoundException, JsonProcessingException {
+        return igdbService.findBySearch(search);
+    }
+
+    @POST
+    @Path("/genres/{genre}")
+    @Produces(MediaType.APPLICATION_JSON)
+    public List<Videogame> getVideogameByGenre(@PathParam("genre") int genre) throws VideogameNotFoundException, JsonProcessingException {
+        return igdbService.findByGenre(genre);
+    }
 }
